Make view init consistent with other views
