+++ Sources/ProcessOutUI/Sources/Modules/DynamicCheckout/View/PODynamicCheckoutView+Init.swift
@@ -1,0 +1,50 @@
+//
+//  PODynamicCheckoutView.swift
+//  ProcessOutUI
+//
+//  Created by Andrii Vysotskyi on 29.02.2024.
+//
+
+@_spi(PO) import ProcessOut
+
+@available(iOS 14, *)
+extension PODynamicCheckoutView {
+
+    /// Creates dynamic checkout view.
+    ///
+    /// - NOTE: Use caution when using this view, because SwiftUI only initializes
+    /// its state once during the lifetime of the view — even if you call the initializer
+    /// more than once — which might result in unexpected behavior.
+    public init(
+        configuration: PODynamicCheckoutConfiguration,
+        delegate: PODynamicCheckoutDelegate,
+        completion: @escaping (Result<Void, POFailure>) -> Void
+    ) {
+        let viewModel = {
+            var logger = ProcessOut.shared.logger
+            logger[attributeKey: "InvoiceId"] = configuration.invoiceId
+            let interactor = DynamicCheckoutDefaultInteractor(
+                configuration: configuration,
+                delegate: delegate,
+                passKitPaymentInteractor: DynamicCheckoutPassKitPaymentDefaultInteractor(
+                    configuration: configuration, delegate: delegate, invoicesService: ProcessOut.shared.invoices
+                ),
+                alternativePaymentInteractor: DynamicCheckoutAlternativePaymentDefaultInteractor(
+                    configuration: configuration.alternativePayment
+                ),
+                childProvider: DefaultDynamicCheckoutInteractorChildProvider(
+                    configuration: configuration,
+                    cardsService: ProcessOut.shared.cards,
+                    invoicesService: ProcessOut.shared.invoices,
+                    imagesRepository: ProcessOut.shared.images,
+                    logger: logger
+                ),
+                invoicesService: ProcessOut.shared.invoices,
+                logger: logger,
+                completion: completion
+            )
+            return DefaultDynamicCheckoutViewModel(interactor: interactor)
+        }
+        self = .init(viewModel: viewModel())
+    }
+}
